LayoutTests/fast/forms/input-widths.html

<!DOCTYPE html>
<html>
<head>
<script src="../../fast/js/resources/js-test-pre.js"></script>
</head>
<body>
<p id="description">Text field types should have the same default widths.</p>
<div id="console"></div>

<div id=parent>
<input type=text id=text>

<input type=email id=email>
<input type=number id=number>
<input type=password id=password>
<input type=tel id=tel>
<input type=url id=url>
</div>

<script>
var baseWidth = document.getElementById('text').offsetWidth;

shouldBe('document.getElementById("email").offsetWidth', 'baseWidth');
shouldBe('document.getElementById("number").offsetWidth', 'baseWidth');
shouldBe('document.getElementById("password").offsetWidth', 'baseWidth');
shouldBe('document.getElementById("tel").offsetWidth', 'baseWidth');
shouldBe('document.getElementById("url").offsetWidth', 'baseWidth');

// type=search is not tested intentionally.

document.getElementById('parent').innerHTML = '';
</script>
<script src="../../fast/js/resources/js-test-post.js"></script>
</body>
</html>
